# Speed up testing