--- a/plugin/forward/health_test.go
+++ b/plugin/forward/health_test.go
@@ -14,7 +14,11 @@ import (
 )
 
 func TestHealth(t *testing.T) {
-	const expected = 1
+	hcReadTimeout = 10 * time.Millisecond
+	hcWriteTimeout = 10 * time.Millisecond
+	readTimeout = 10 * time.Millisecond
+	defaultTimeout = 10 * time.Millisecond
+
 	i := uint32(0)
 	q := uint32(0)
 	s := dnstest.NewServer(func(w dns.ResponseWriter, r *dns.Msg) {
@@ -41,15 +45,19 @@ func TestHealth(t *testing.T) {
 
 	f.ServeDNS(context.TODO(), &test.ResponseWriter{}, req)
 
-	time.Sleep(1 * time.Second)
+	time.Sleep(20 * time.Millisecond)
 	i1 := atomic.LoadUint32(&i)
-	if i1 != expected {
-		t.Errorf("Expected number of health checks with RecursionDesired==true to be %d, got %d", expected, i1)
+	if i1 != 1 {
+		t.Errorf("Expected number of health checks with RecursionDesired==true to be %d, got %d", 1, i1)
 	}
 }
 
 func TestHealthNoRecursion(t *testing.T) {
-	const expected = 1
+	hcReadTimeout = 10 * time.Millisecond
+	readTimeout = 10 * time.Millisecond
+	defaultTimeout = 10 * time.Millisecond
+	hcWriteTimeout = 10 * time.Millisecond
+
 	i := uint32(0)
 	q := uint32(0)
 	s := dnstest.NewServer(func(w dns.ResponseWriter, r *dns.Msg) {
@@ -77,15 +85,19 @@ func TestHealthNoRecursion(t *testing.T) {
 
 	f.ServeDNS(context.TODO(), &test.ResponseWriter{}, req)
 
-	time.Sleep(1 * time.Second)
+	time.Sleep(20 * time.Millisecond)
 	i1 := atomic.LoadUint32(&i)
-	if i1 != expected {
-		t.Errorf("Expected number of health checks with RecursionDesired==false to be %d, got %d", expected, i1)
+	if i1 != 1 {
+		t.Errorf("Expected number of health checks with RecursionDesired==false to be %d, got %d", 1, i1)
 	}
 }
 
 func TestHealthTimeout(t *testing.T) {
-	const expected = 1
+	hcReadTimeout = 10 * time.Millisecond
+	hcWriteTimeout = 10 * time.Millisecond
+	readTimeout = 10 * time.Millisecond
+	defaultTimeout = 10 * time.Millisecond
+
 	i := uint32(0)
 	q := uint32(0)
 	s := dnstest.NewServer(func(w dns.ResponseWriter, r *dns.Msg) {
@@ -117,58 +129,20 @@ func TestHealthTimeout(t *testing.T) {
 
 	f.ServeDNS(context.TODO(), &test.ResponseWriter{}, req)
 
-	time.Sleep(1 * time.Second)
+	time.Sleep(20 * time.Millisecond)
 	i1 := atomic.LoadUint32(&i)
-	if i1 != expected {
-		t.Errorf("Expected number of health checks to be %d, got %d", expected, i1)
-	}
-}
-
-func TestHealthFailTwice(t *testing.T) {
-	const expected = 2
-	i := uint32(0)
-	q := uint32(0)
-	s := dnstest.NewServer(func(w dns.ResponseWriter, r *dns.Msg) {
-		if r.Question[0].Name == "." {
-			atomic.AddUint32(&i, 1)
-			i1 := atomic.LoadUint32(&i)
-			// Timeout health until we get the second one
-			if i1 < 2 {
-				return
-			}
-			ret := new(dns.Msg)
-			ret.SetReply(r)
-			w.WriteMsg(ret)
-			return
-		}
-		if atomic.LoadUint32(&q) == 0 { //drop only first query
-			atomic.AddUint32(&q, 1)
-			return
-		}
-		ret := new(dns.Msg)
-		ret.SetReply(r)
-		w.WriteMsg(ret)
-	})
-	defer s.Close()
-
-	p := NewProxy(s.Addr, transport.DNS)
-	f := New()
-	f.SetProxy(p)
-	defer f.OnShutdown()
-
-	req := new(dns.Msg)
-	req.SetQuestion("example.org.", dns.TypeA)
-
-	f.ServeDNS(context.TODO(), &test.ResponseWriter{}, req)
-
-	time.Sleep(3 * time.Second)
-	i1 := atomic.LoadUint32(&i)
-	if i1 != expected {
-		t.Errorf("Expected number of health checks to be %d, got %d", expected, i1)
+	if i1 != 1 {
+		t.Errorf("Expected number of health checks to be %d, got %d", 1, i1)
 	}
 }
 
 func TestHealthMaxFails(t *testing.T) {
+	hcReadTimeout = 10 * time.Millisecond
+	hcWriteTimeout = 10 * time.Millisecond
+	readTimeout = 10 * time.Millisecond
+	defaultTimeout = 10 * time.Millisecond
+	hcInterval = 10 * time.Millisecond
+
 	s := dnstest.NewServer(func(w dns.ResponseWriter, r *dns.Msg) {
 		// timeout
 	})
@@ -185,7 +159,7 @@ func TestHealthMaxFails(t *testing.T) {
 
 	f.ServeDNS(context.TODO(), &test.ResponseWriter{}, req)
 
-	time.Sleep(readTimeout + 1*time.Second)
+	time.Sleep(100 * time.Millisecond)
 	fails := atomic.LoadUint32(&p.fails)
 	if !p.Down(f.maxfails) {
 		t.Errorf("Expected Proxy fails to be greater than %d, got %d", f.maxfails, fails)
@@ -193,7 +167,12 @@ func TestHealthMaxFails(t *testing.T) {
 }
 
 func TestHealthNoMaxFails(t *testing.T) {
-	const expected = 0
+	hcReadTimeout = 10 * time.Millisecond
+	hcWriteTimeout = 10 * time.Millisecond
+	readTimeout = 10 * time.Millisecond
+	defaultTimeout = 10 * time.Millisecond
+	hcInterval = 10 * time.Millisecond
+
 	i := uint32(0)
 	s := dnstest.NewServer(func(w dns.ResponseWriter, r *dns.Msg) {
 		if r.Question[0].Name == "." {
@@ -217,9 +196,9 @@ func TestHealthNoMaxFails(t *testing.T) {
 
 	f.ServeDNS(context.TODO(), &test.ResponseWriter{}, req)
 
-	time.Sleep(1 * time.Second)
+	time.Sleep(20 * time.Millisecond)
 	i1 := atomic.LoadUint32(&i)
-	if i1 != expected {
-		t.Errorf("Expected number of health checks to be %d, got %d", expected, i1)
+	if i1 != 0 {
+		t.Errorf("Expected number of health checks to be %d, got %d", 0, i1)
 	}
 }