--- a/plugin/forward/persistent.go
+++ b/plugin/forward/persistent.go
@@ -152,7 +152,10 @@ const (
 	defaultExpire  = 10 * time.Second
 	minDialTimeout = 1 * time.Second
 	maxDialTimeout = 30 * time.Second
+)
 
+// Make a var for minimizing this value in tests.
+var (
 	// Some resolves might take quite a while, usually (cached) responses are fast. Set to 2s to give us some time to retry a different upstream.
 	readTimeout = 2 * time.Second
 )