--- a/plugin/metrics/metrics_test.go
+++ b/plugin/metrics/metrics_test.go
@@ -30,25 +30,25 @@ func TestMetrics(t *testing.T) {
 		// This all works because 1 bucket (1 zone, 1 type)
 		{
 			next:          test.NextHandler(dns.RcodeSuccess, nil),
-			qname:         "example.org",
+			qname:         "example.org.",
 			metric:        "coredns_dns_requests_total",
 			expectedValue: "1",
 		},
 		{
 			next:          test.NextHandler(dns.RcodeSuccess, nil),
-			qname:         "example.org",
+			qname:         "example.org.",
 			metric:        "coredns_dns_requests_total",
 			expectedValue: "2",
 		},
 		{
 			next:          test.NextHandler(dns.RcodeSuccess, nil),
-			qname:         "example.org",
+			qname:         "example.org.",
 			metric:        "coredns_dns_requests_total",
 			expectedValue: "3",
 		},
 		{
 			next:          test.NextHandler(dns.RcodeSuccess, nil),
-			qname:         "example.org",
+			qname:         "example.org.",
 			metric:        "coredns_dns_responses_total",
 			expectedValue: "4",
 		},
@@ -61,7 +61,7 @@ func TestMetrics(t *testing.T) {
 		if tc.qtype == 0 {
 			tc.qtype = dns.TypeA
 		}
-		req.SetQuestion(dns.Fqdn(tc.qname), tc.qtype)
+		req.SetQuestion(tc.qname, tc.qtype)
 		met.Next = tc.next
 
 		rec := dnstest.NewRecorder(&test.ResponseWriter{})