--- a/plugin/sign/signer_test.go
+++ b/plugin/sign/signer_test.go
@@ -1,7 +1,6 @@
 package sign
 
 import (
-	"bytes"
 	"io/ioutil"
 	"os"
 	"testing"
@@ -176,8 +175,4 @@ func TestSignDS(t *testing.T) {
 	if x := nsec[0].Header().Ttl; x != minttl {
 		t.Errorf("Expected no NSEC TTL to be %d for %s, got %d", minttl, "www.miek.nl.", x)
 	}
-	// print zone on error
-	buf := &bytes.Buffer{}
-	write(buf, z)
-	t.Logf("%s\n", buf)
 }