--- a/test/auto_test.go
+++ b/test/auto_test.go
@@ -21,7 +21,7 @@ func TestAuto(t *testing.T) {
 	corefile := `org:0 {
 		auto {
 			directory ` + tmpdir + ` db\.(.*) {1}
-			reload 0.1s
+			reload 0.01s
 		}
 	}`
 
@@ -46,7 +46,7 @@ func TestAuto(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	time.Sleep(150 * time.Millisecond) // wait for it to be picked up
+	time.Sleep(10 * time.Millisecond) // wait for it to be picked up
 
 	resp, err = dns.Exchange(m, udp)
 	if err != nil {
@@ -59,7 +59,7 @@ func TestAuto(t *testing.T) {
 	// Remove db.example.org again.
 	os.Remove(filepath.Join(tmpdir, "db.example.org"))
 
-	time.Sleep(150 * time.Millisecond) // wait for it to be picked up
+	time.Sleep(10 * time.Millisecond) // wait for it to be picked up
 	resp, err = dns.Exchange(m, udp)
 	if err != nil {
 		t.Fatal("Expected to receive reply, but didn't")
@@ -80,7 +80,7 @@ func TestAutoNonExistentZone(t *testing.T) {
 	corefile := `.:0 {
 		auto {
 			directory ` + tmpdir + ` (.*) {1}
-			reload 1s
+			reload 0.1s
 		}
 		errors stdout
 	}`
@@ -119,7 +119,7 @@ func TestAutoAXFR(t *testing.T) {
 	corefile := `org:0 {
 		auto {
 			directory ` + tmpdir + ` db\.(.*) {1}
-			reload 0.1s
+			reload 0.01s
 		}
 		transfer {
 			to *
@@ -142,7 +142,7 @@ func TestAutoAXFR(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	time.Sleep(150 * time.Millisecond) // wait for it to be picked up
+	time.Sleep(10 * time.Millisecond) // wait for it to be picked up
 
 	tr := new(dns.Transfer)
 	m := new(dns.Msg)