--- a/test/proxy_health_test.go
+++ b/test/proxy_health_test.go
@@ -2,6 +2,7 @@ package test
 
 import (
 	"testing"
+	"time"
 
 	"github.com/miekg/dns"
 )
@@ -61,9 +62,11 @@ func TestProxyThreeWay(t *testing.T) {
 
 	m := new(dns.Msg)
 	m.SetQuestion("example.org.", dns.TypeA)
+	c := new(dns.Client)
+	c.Timeout = 10 * time.Millisecond
 
 	for i := 0; i < 10; i++ {
-		r, err := dns.Exchange(m, addr)
+		r, _, err := c.Exchange(m, addr)
 		if err != nil {
 			continue
 		}