@miekg 2020-10-26
Another 10s can be shaved off as we fix https://github.com/coredns/coredns/issues/4238, meaning all tests run 5s which looks like nice ballpark figure to make local dev. fun again