@miekg 2020-10-26
let me fix the timing as I took the wrong value from `time`