@miekg 2020-10-26
travis is becoming very very slow