@miekg 2020-10-26
yes, would also cut down on external deps, circle-ci is already finished, and that does a minikube k8s test!